refactor: split ServerSpec into two
