+++ zio-http/src/test/scala/zhttp/service/ServerSpec.scala
@@ -23,23 +23,7 @@ object ServerSpec extends HttpRunnableSpec {
   private val env =
     EventLoopGroup.nio() ++ ChannelFactory.nio ++ ServerChannelFactory.nio ++ DynamicServer.live
 
-  private val staticApp = Http.collectZIO[Request] {
-    case Method.GET -> !! / "success"       => ZIO.succeed(Response.ok)
-    case Method.GET -> !! / "failure"       => ZIO.fail(new RuntimeException("FAILURE"))
-    case Method.GET -> !! / "get%2Fsuccess" => ZIO.succeed(Response.ok)
-  }
-
-  private val throwableApp: Http[Any, Nothing, Request, Nothing] = Http.collectZIO[Request] {
-    case Method.GET -> !! / "throwable" => throw new Exception("Throw inside Handler")
-  }
-
-  // Use this route to test anything that doesn't require ZIO related computations.
-  private val nonZIO = Http.collectHExit[Request] {
-    case _ -> !! / "HExitSuccess" => HExit.succeed(Response.ok)
-    case _ -> !! / "HExitFailure" => HExit.fail(new RuntimeException("FAILURE"))
-  }
-
-  private val app = serve { nonZIO ++ staticApp ++ throwableApp ++ DynamicServer.app }
+  private val app = serve { DynamicServer.app }
 
   def dynamicAppSpec = suite("DynamicAppSpec") {
     suite("success") {
@@ -117,40 +101,6 @@ object ServerSpec extends HttpRunnableSpec {
           assertM(res)(equalTo("abc"))
         }
       }
-  }
-
-  def nonZIOSpec = suite("NonZIOSpec") {
-    testM("200 response") {
-      checkAllM(HttpGen.method) { method =>
-        val actual = status(method, !! / "HExitSuccess")
-        assertM(actual)(equalTo(Status.OK))
-      }
-    } +
-      testM("500 response") {
-        val methodGenWithoutHEAD: Gen[Any, Method] = Gen.fromIterable(
-          List(
-            Method.OPTIONS,
-            Method.GET,
-            Method.POST,
-            Method.PUT,
-            Method.PATCH,
-            Method.DELETE,
-            Method.TRACE,
-            Method.CONNECT,
-          ),
-        )
-        checkAllM(methodGenWithoutHEAD) { method =>
-          val actual = status(method, !! / "HExitFailure")
-          assertM(actual)(equalTo(Status.INTERNAL_SERVER_ERROR))
-        }
-      } +
-      testM("404 response ") {
-        checkAllM(HttpGen.method) { method =>
-          val actual = status(method, !! / "A")
-          assertM(actual)(equalTo(Status.NOT_FOUND))
-        }
-      }
-
   }
 
   def requestSpec = suite("RequestSpec") {
@@ -265,67 +215,13 @@ object ServerSpec extends HttpRunnableSpec {
       }
   }
 
-  def serverStartSpec = suite("ServerStartSpec") {
-    testM("desired port") {
-      val port = 8088
-      (Server.port(port) ++ Server.app(Http.empty)).make.use { start =>
-        assertM(ZIO.effect(start.port))(equalTo(port))
-      }
-    } +
-      testM("available port") {
-        (Server.port(0) ++ Server.app(Http.empty)).make.use { start =>
-          assertM(ZIO.effect(start.port))(not(equalTo(0)))
-        }
-      }
-  }
-
   override def spec =
     suiteM("Server") {
       app
         .as(
-          List(serverStartSpec, staticAppSpec, dynamicAppSpec, responseSpec, requestSpec, nonZIOSpec, throwableAppSpec),
+          List(dynamicAppSpec, responseSpec, requestSpec),
         )
         .useNow
     }.provideCustomLayerShared(env) @@ timeout(30 seconds)
 
-  def staticAppSpec    = suite("StaticAppSpec") {
-    testM("200 response") {
-      val actual = status(path = !! / "success")
-      assertM(actual)(equalTo(Status.OK))
-    } +
-      testM("500 response") {
-        val actual = status(path = !! / "failure")
-        assertM(actual)(equalTo(Status.INTERNAL_SERVER_ERROR))
-      } +
-      testM("404 response") {
-        val actual = status(path = !! / "random")
-        assertM(actual)(equalTo(Status.NOT_FOUND))
-      } +
-      testM("200 response with encoded path") {
-        val actual = status(path = !! / "get%2Fsuccess")
-        assertM(actual)(equalTo(Status.OK))
-      } +
-      testM("Multiple 200 response") {
-        for {
-          data <- status(path = !! / "success").repeatN(1024)
-        } yield assertTrue(data == Status.OK)
-      }
-  }
-  def throwableAppSpec = suite("ThrowableAppSpec") {
-    testM("Throw inside Handler") {
-      for {
-        status <- status(Method.GET, !! / "throwable")
-      } yield assertTrue(status == Status.INTERNAL_SERVER_ERROR)
-    } + testM("Throw inside Dynamic App") {
-      for {
-        status <- Http
-          .collectZIO[Request] { case Method.GET -> !! / "throwableD" =>
-            throw new Exception("Throw inside Dynamic App")
-          }
-          .deploy
-          .status
-          .run(!! / "throwableD")
-      } yield assertTrue(status == Status.INTERNAL_SERVER_ERROR)
-    }
-  }
 }
